requirements/other.txt
celery
django-celery-results
selenium
